Add files via upload
In this project, I queried data from a mental health survey. The data in question focused on university students (undergraduate and postgraduate) and their ratings on their mental health along with other questions such as activities for stress relief and future insecurity. This project explored answered my questions on the dataset.